honeybox/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.10.0
python-multipart>=0.0.6
feedparser>=6.0.10
python-dateutil>=2.8.2
requests>=2.31.0
boto3>=1.34.0 